django/requirements.txt
Django==1.11.7
django-compressor==2.2
django-loginas==0.3.2
djangorestframework==3.8
django-npm==1.0.0
psycopg2-binary==2.8.5
drf-yasg==1.17.0
git+https://github.com/harpb/pypugjs.git@master
django_elasticsearch_dsl==7.1.4